fixed run test with anonymous class by php early 7.0
