+++ tests/unit/QA/SoftMocksTestClasses.php
@@ -132,21 +132,6 @@ class GrandChildStaticTestClass extends ChildStaticTestClass
     }
 }
 
-class AnonymousTestClass
-{
-    const SOMETHING = 1;
-
-    public function doSomething()
-    {
-        return new class {
-            public function meth()
-            {
-                return [self::class, static::class, get_called_class()];
-            }
-        };
-    }
-}
-
 class WithExitTestClass
 {
     const RESULT = 42;
